requirements.txt
amqp==1.4.6
anyjson==0.3.3
backports.ssl-match-hostname==3.4.0.2
billiard==3.3.0.19
celery==3.1.17
certifi==14.5.14
Django==1.8.4
django-async-messages==0.3.1
django-celery==3.1.16
django-cors-headers==1.1.0
django-password-reset==0.7
django-resized==0.2.8
djangorestframework==3.0.3
flower==0.7.3
kombu==3.0.24
Pillow==2.7.0
pytz==2014.10
redis==2.10.3
six==1.9.0
sorl-thumbnail==12.2
tornado==4.1
dj-database-url==0.3.0
dj-static==0.0.6
gunicorn==19.3.0
static3==0.6.1
psycopg2==2.6.1